Respect color option when instantiating shells
Thor's base shell will be memoized the first time it is set. So further
instantiations of a bundler shell will be initialized with a color
shell, even if no-color is specified. This is causing some sources specs to
fail, depending on the order they run.

See for example https://travis-ci.org/bundler/bundler/builds/500328994.